Renamed PreceptorEvaluation model to following naming guidelines
