+++ server/src/controllers/PreceptorController.js
@@ -1,4 +1,4 @@
-import PEval from '../models/PreceptorEvaluationModel.js';
+import PEval from '../models/PreceptorEvaluation.js';
 
 //get evals that match a preceptor ID
 export const getEvals = async (req, res) => {
@@ -42,12 +42,10 @@ export async function getEval(req, res) {
 export async function editEval(req, res) {
     const id = req.body._id;
 
-  try{
-    await PEval.findByIdAndUpdate(
-      {_id: id}, {...req.body}
-      )
-  } catch(err){
-    console.log(res.json({msg: err.message}));
-  }
-  return res.status(200).json({test: 'test'});
+    try {
+        await PEval.findByIdAndUpdate({ _id: id }, { ...req.body });
+    } catch (err) {
+        console.log(res.json({ msg: err.message }));
+    }
+    return res.status(200).json({ test: 'test' });
 }
